WIP: Submit esp-mqtt sub/unsub/pub requests asynchronously
The ESP-IDF MQTT component is fairly unusable for low-latency setups such
as ESPHome. By default, ESPHome calls esp_mqtt_client_publish() directly
from the MQTT component's loop() function, or when publishing status
updates for other components. This may block for up to 20 seconds(!!) in
adverse network conditions.

With the `idf_send_async` option, subscribe and unsubscribe requests can
still block the loop thread for multiple seconds, but publishing sensor
updates is queued for esp-mqtt's own thread to actually send them. Which
it does very slowly, no more than one per second, as discussed in
https://github.com/espressif/esp-idf/issues/15458

And to top it all off, even with `idf_send_async` set, the so-called
'asynchronous' send can still block for ten seconds because it takes
the same MQTT_API_LOCK that the esp-mqtt thread holds while it runs
its loop and a network send is timing out. This is reported in
https://github.com/espressif/esp-idf/issues/13078

The only way I can see to use esp-mqtt sanely is to use a thread of our
own, queueing all sub/unsub/publish requests and invoking the esp-mqtt
APIs from that thread.

The existing RingBuffer abstraction works nicely for this as it already
handles all the atomicity and waking when data are available. I've chosen
to avoid allocations by passing the actual data through the ringbuffer,
which means we impose a hard limit on the total topic+payload size for
each request. An alternative would be to allocate copies in the enqueue()
function and to pass *pointers* through the ringbuffer (which could be
a different type of queue then, if we wanted to reinvent things).

Fixes: #6810