Initialise h-bridge switch to requested initial state
Fixes https://github.com/esphome/issues/issues/6833

The value_or(false) call converted the optional<bool> to a plain bool and write_state(initial_state) was writing whether or not there was initial state (which there always would be because of the value_or call). This meant that regardless of the requested restore_mode in the config, the switch would always be turned on at boot.