real_time_clock: Apply timezone immediately in set_timezone()
The timezone is handled by the C library, when ESPHome sets the $TZ
environment variable and then calls tzset().

That's done from the ->apply_timezone() method which is currently
invoked only from ->call_setup(). For the SNTP component, setup is
actually quite late in boot (setup_priority::BEFORE_CONNECTION).

So everything which runs *before* that will get incorrect times in GMT
instead of localtime, because the timezone hasn't been set up yet. That
includes on_boot methods at the default priority as well as a lot of
other setup.

Fix that by telling the C library about the timezone as early as
possible, directly in the ->set_timezone() method.

Closes https://github.com/esphome/issues/issues/6840