Fix '--device MQTT' for devices with static IP
An earlier commit disabled the MQTT IP lookup if the device has a static
IP address in CORE.address. That made sense for the automatic fallback
to MQTT when mDNS is disabled, but not if the user explicitly asked for
MQTT.

Rework the condition to apply that check only to the !mDNS case, and clean
it up to be a little more understandable.

Fixes: ac0549578197 ("Dont do mqtt ip lookup if `use_address` has ip address (#5096)")