OTA: Close and clean up client when setsockopt fails
When setsockopt call fails, for example on ECONNRESET, the client
connection never gets cleaned up properly. This causes a client that
connects and disconnects right away to cause the OTA to be stuck in
a loop and logspam with error. This change ensures when setsockopt
call fails, the client gets cleaned up properly.