Update sm300d2.cpp
This has been a complete rewrite of how the data is parsed. 

This now receives the data strings and parses out both the original 17 octet and the newer 19 octet data strings. 

It places these into a receive buffer and then pulls the data out.
Also can receive partial data strings for assembly when the rest is received. 

 Once it see a timestamp update, it will clear the buffer to ensure no stale date is kept.

-Enjoy