Add Lightning LN882H support
Libretiny added support for Lightning LN882H in 1.9.0:
https://github.com/libretiny-eu/libretiny/releases/tag/v1.9.0

This change adds the missing pieces in esphome to allow building for
the new ln822x platform.

https://docs.libretiny.eu/docs/platform/lightning-ln882x/
https://github.com/libretiny-eu/ltchiptool/commit/c85d5e0b2bd4639ffcc0543f46ae1f15395606c6
https://github.com/libretiny-eu/framework-lightning-ln882h

The board definition was generated by the generate_components.py script
based on the families.json info and board information from the Libretiny
repository as was previously done for bk72xx and rtl87xx.

The rest of modifications are mostly constant additions and inclusions
analogous to the ones previously made for other chips supported by
Libretiny.